+++ extensions/reddit/src/main/java/app/revanced/extension/reddit/patches/FilterPromotedLinksPatch.java
@@ -1,12 +1,16 @@
-package app.revanced.extension.patches;
+package app.revanced.extension.reddit.patches;
 
 import com.reddit.domain.model.ILink;
 
 import java.util.ArrayList;
 import java.util.List;
 
+@SuppressWarnings("unused")
 public final class FilterPromotedLinksPatch {
+
     /**
+     * Injection point.
+     *
      * Filters list from promoted links.
      **/
     public static List<?> filterChildren(final Iterable<?> links) {